working

+++ src/server/Models/NeuralItemResult.cs
@@ -2,7 +2,7 @@ namespace Models
 {
     public class NeuralItemResult
     {
-        public int NeuralOrder { get; set; }
+        public double NeuralOrder { get; set; }
         public int LeadInQuery { get; set; }
         public Offer Item { get; set; }
     }
